backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2==2.9.9
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uvicorn==0.29.0

uuid~=1.30
transformers~=4.39.3
pillow~=10.3.0
torch~=2.2.2
python-dotenv~=1.0.1
PyYAML~=6.0.1